// src/models/battle/buff.enum.ts
export enum BuffEnum {
  BURN = 1,
  WET = 2,
  BLIND = 3,
  STATIC = 4,
  STUN = 5,
  BOOST = 6,
  WELTGEIST = 7,
  BARRIER = 8,
  RAGE = 9,
  AGGRO = 10,
  COUNTER = 11,
  POWER_DOWN = 12,
  DEFENSE_DOWN = 13,
  SPEED_UP = 14,
  SPEED_DOWN = 15,
  DEFENSE_UP = 16,
  SILENCE = 17,
  POWER_UP = 18
}